Moving design system references to shared version.
