+++ android/app/src/androidTest/java/com/adammcneilly/pocketleague/Utils.kt
@@ -2,7 +2,7 @@ package com.adammcneilly.pocketleague
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
-import com.adammcneilly.pocketleague.android.designsystem.theme.PocketLeagueTheme
+import com.adammcneilly.pocketleague.shared.design.system.theme.PocketLeagueTheme
 
 /**
  * Helper function that renders the supplied [content] inside a [PocketLeagueTheme].
