+++ android/app/src/main/java/com/adammcneilly/pocketleague/MainActivity.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.ui.graphics.Color
 import androidx.core.view.WindowCompat
-import com.adammcneilly.pocketleague.android.designsystem.theme.PocketLeagueTheme
+import com.adammcneilly.pocketleague.shared.design.system.theme.PocketLeagueTheme
 import com.adammcneilly.pocketleague.ui.MainComposable
 import com.adammcneilly.pocketleague.ui.sizeconfigs.ContentType
 import com.adammcneilly.pocketleague.ui.sizeconfigs.NavigationType
