+++ android/app/src/main/java/com/adammcneilly/pocketleague/ui/MatchDetail.kt
@@ -20,10 +20,10 @@ import com.adammcneilly.pocketleague.android.designsystem.game.GameListItem
 import com.adammcneilly.pocketleague.android.designsystem.matches.MatchDetailHeader
 import com.adammcneilly.pocketleague.android.designsystem.players.TeamRosterCard
 import com.adammcneilly.pocketleague.android.designsystem.stats.CoreStatsComparisonCard
-import com.adammcneilly.pocketleague.android.designsystem.theme.rlcsBlue
-import com.adammcneilly.pocketleague.android.designsystem.theme.rlcsOrange
 import com.adammcneilly.pocketleague.core.displaymodels.GameDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
+import com.adammcneilly.pocketleague.shared.design.system.theme.rlcsBlue
+import com.adammcneilly.pocketleague.shared.design.system.theme.rlcsOrange
 
 /**
  * Renders the [displayModel] to show detailed information about a match between two teams.
