+++ android/design-system/src/main/java/com/adammcneilly/pocketleague/android/designsystem/stats/CoreStatsComparisonCard.kt
@@ -11,8 +11,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.adammcneilly.pocketleague.android.designsystem.theme.PocketLeagueTheme
 import com.adammcneilly.pocketleague.core.displaymodels.CoreStatsDisplayModel
+import com.adammcneilly.pocketleague.shared.design.system.theme.PocketLeagueTheme
 
 /**
  * Render a comparison of [blueTeamStats] and [orangeTeamStats] for the
