--- android/design-system/src/main/java/com/adammcneilly/pocketleague/android/designsystem/theme/Color.kt
@@ -1,67 +1,0 @@
-@file:Suppress("MagicNumber")
-
-package com.adammcneilly.pocketleague.android.designsystem.theme
-
-import androidx.compose.ui.graphics.Color
-
-val md_theme_light_primary = Color(0xFF3d5ba9)
-val md_theme_light_onPrimary = Color(0xFFffffff)
-val md_theme_light_primaryContainer = Color(0xFFdae2ff)
-val md_theme_light_onPrimaryContainer = Color(0xFF00174d)
-val md_theme_light_secondary = Color(0xFF585e71)
-val md_theme_light_onSecondary = Color(0xFFffffff)
-val md_theme_light_secondaryContainer = Color(0xFFdce1f9)
-val md_theme_light_onSecondaryContainer = Color(0xFF151b2c)
-val md_theme_light_tertiary = Color(0xFF735470)
-val md_theme_light_onTertiary = Color(0xFFffffff)
-val md_theme_light_tertiaryContainer = Color(0xFFffd6f9)
-val md_theme_light_onTertiaryContainer = Color(0xFF2b122b)
-val md_theme_light_error = Color(0xFFba1b1b)
-val md_theme_light_errorContainer = Color(0xFFffdad4)
-val md_theme_light_onError = Color(0xFFffffff)
-val md_theme_light_onErrorContainer = Color(0xFF410001)
-val md_theme_light_background = Color(0xFFfefbff)
-val md_theme_light_onBackground = Color(0xFF1b1b1f)
-val md_theme_light_surface = Color(0xFFfefbff)
-val md_theme_light_onSurface = Color(0xFF1b1b1f)
-val md_theme_light_surfaceVariant = Color(0xFFe2e2ec)
-val md_theme_light_onSurfaceVariant = Color(0xFF44464e)
-val md_theme_light_outline = Color(0xFF75767f)
-val md_theme_light_inverseOnSurface = Color(0xFFf2f0f5)
-val md_theme_light_inverseSurface = Color(0xFF303033)
-val md_theme_light_inversePrimary = Color(0xFFb1c5ff)
-val md_theme_light_shadow = Color(0xFF000000)
-
-val md_theme_dark_primary = Color(0xFFb1c5ff)
-val md_theme_dark_onPrimary = Color(0xFF002a78)
-val md_theme_dark_primaryContainer = Color(0xFF214290)
-val md_theme_dark_onPrimaryContainer = Color(0xFFdae2ff)
-val md_theme_dark_secondary = Color(0xFFc1c6dd)
-val md_theme_dark_onSecondary = Color(0xFF2a3042)
-val md_theme_dark_secondaryContainer = Color(0xFF414659)
-val md_theme_dark_onSecondaryContainer = Color(0xFFdce1f9)
-val md_theme_dark_tertiary = Color(0xFFe1bbdc)
-val md_theme_dark_onTertiary = Color(0xFF412740)
-val md_theme_dark_tertiaryContainer = Color(0xFF5a3d58)
-val md_theme_dark_onTertiaryContainer = Color(0xFFffd6f9)
-val md_theme_dark_error = Color(0xFFffb4a9)
-val md_theme_dark_errorContainer = Color(0xFF930006)
-val md_theme_dark_onError = Color(0xFF680003)
-val md_theme_dark_onErrorContainer = Color(0xFFffdad4)
-val md_theme_dark_background = Color(0xFF1b1b1f)
-val md_theme_dark_onBackground = Color(0xFFe3e1e6)
-val md_theme_dark_surface = Color(0xFF1b1b1f)
-val md_theme_dark_onSurface = Color(0xFFe3e1e6)
-val md_theme_dark_surfaceVariant = Color(0xFF44464e)
-val md_theme_dark_onSurfaceVariant = Color(0xFFc6c6d0)
-val md_theme_dark_outline = Color(0xFF8f909a)
-val md_theme_dark_inverseOnSurface = Color(0xFF1b1b1f)
-val md_theme_dark_inverseSurface = Color(0xFFe3e1e6)
-val md_theme_dark_inversePrimary = Color(0xFF3d5ba9)
-val md_theme_dark_shadow = Color(0xFF000000)
-
-val rlcsBlue = Color(0xFF0069c1)
-val rlcsOrange = Color(0xFFee7d13)
-
-val seed = Color(0xFF3d5ba9)
-val error = Color(0xFFba1b1b)
