+++ gradle/libs.versions.toml
@@ -10,7 +10,7 @@ buildkonfig = "0.13.3"
 coil = "2.4.0"
 compileSdk = "33"
 compose = "2023.05.01"
-composeCompiler = "1.4.7"
+composeCompiler = "1.4.6"
 composeMaterial3 = "1.0.1"
 espresso = "3.5.1"
 glance = "1.0.0-alpha05"
